iommu/sva: Assign a PASID to mm on PASID allocation and free it on mm exit
mainline inclusion
from mainline-v5.18-rc1
commit bb851179d340e4779a498dc9380a9fafdb726dfd
category: feature
bugzilla: https://gitee.com/openeuler/intel-kernel/issues/I590PB
CVE: NA

Intel-SIG: commit bb851179d340 iommu/sva: Assign a PASID to mm on
PASID allocation and free it on mm exit
Backport for intel SPR ENQCMD Instruction support

--------------------------------

PASIDs are process-wide. It was attempted to use refcounted PASIDs to
free them when the last thread drops the refcount. This turned out to
be complex and error prone. Given the fact that the PASID space is 20
bits, which allows up to 1M processes to have a PASID associated
concurrently, PASID resource exhaustion is not a realistic concern.

Therefore, it was decided to simplify the approach and stick with lazy
on demand PASID allocation, but drop the eager free approach and make an
allocated PASID's lifetime bound to the lifetime of the process.

Get rid of the refcounting mechanisms and replace/rename the interfaces
to reflect this new approach.

  [ bp: Massage commit message. ]

Suggested-by: Dave Hansen <dave.hansen@linux.intel.com>
Signed-off-by: Fenghua Yu <fenghua.yu@intel.com>
Signed-off-by: Borislav Petkov <bp@suse.de>
Reviewed-by: Tony Luck <tony.luck@intel.com>
Reviewed-by: Lu Baolu <baolu.lu@linux.intel.com>
Reviewed-by: Jacob Pan <jacob.jun.pan@linux.intel.com>
Reviewed-by: Thomas Gleixner <tglx@linutronix.de>
Acked-by: Joerg Roedel <jroedel@suse.de>
Link: https://lore.kernel.org/r/20220207230254.3342514-6-fenghua.yu@intel.com
Signed-off-by: Jun Yang <jun.yang@suse.com>